backend/admin/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
django-cors-headers==4.0.0
python-decouple==3.8
cryptography==41.0.0